code/llava/requirements.txt
absl-py==1.4.0
accelerate==0.17.1
aiofiles==22.1.0
aiohttp==3.8.3
aiosignal==1.3.1
aiosqlite==0.18.0
altair==4.2.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.2
bitsandbytes==0.36.0.post2
bleach==6.0.0
blis==0.7.9
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
comm==0.1.2
confection==0.0.4
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.7
datasets==2.9.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
docstring-parser==0.15
einops==0.6.0
emoji==2.2.0
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.5.0/en_core_web_md-3.5.0-py3-none-any.whl
entrypoints==0.4
exceptiongroup==1.1.0
executing==1.2.0
faiss-cpu==1.7.3
fastapi==0.89.1
fastjsonschema==2.16.2
ffmpy==0.3.0
filelock==3.9.0
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2022.11.0
gdown==4.6.0
google-api-core==2.11.0
google-api-python-client==2.77.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.58.0
gradio==3.16.2
greenlet==1.1.0
h11==0.14.0
httpcore==0.16.3
httplib2==0.21.0
httpx==0.23.3
huggingface-hub==0.11.1
idna==3.4
importlib-metadata==6.0.0
iniconfig==2.0.0
ipdb==0.13.11
ipykernel==6.21.2
ipython==8.8.0
ipython-genutils==0.2.0
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.5.0
jupyter-ydoc==0.2.2
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyter_server==2.2.1
jupyter_server_fileid==0.6.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.6.1
jupyterlab==3.6.1
jupyterlab-pygments==0.2.2
jupyterlab_server==2.19.0
kiwisolver==1.4.4
langcodes==3.3.0
linkify-it-py==1.0.3
markdown-it-py==2.1.0
MarkupSafe==2.1.1
matplotlib==3.6.3
matplotlib-inline==0.1.6
mauve-text==0.3.0
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==2.0.5
msgpack==1.0.0
multidict==6.0.4
multiprocess==0.70.14
munch==2.5.0
murmurhash==1.0.9
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
nltk==3.8.1
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
orjson==3.8.5
packaging==23.0
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
pathy==0.10.1
peft==0.2.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.0.0
pluggy==1.0.0
portalocker==2.7.0
preshed==3.0.8
progressbar==2.5
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==4.21.12
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocoevalcap==1.2
pycocotools==2.0.6
pycparser==2.21
pycryptodome==3.16.0
pydantic==1.10.4
pydub==0.25.1
Pygments==2.14.0
pynvim==0.4.3
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
pytest==7.2.1
python-dateutil==2.8.2
python-json-logger==2.0.5
python-multipart==0.0.5
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rouge==1.0.1
rsa==4.9
sacrebleu==1.4.10
sacremoses==0.0.53
scikit-learn==1.2.1
scipy==1.5.4
seaborn==0.12.2
Send2Trash==1.8.0
sentencepiece==0.1.97
simcse==0.4
simctg==0.7
simple-parsing==0.0.21.post1
six==1.16.0
sklearn==0.0.post1
smart-open==6.3.0
sniffio==1.3.0
soupsieve==2.3.2.post1
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.5
stack-data==0.6.2
stanza==1.4.2
starlette==0.22.0
terminado==0.17.1
thinc==8.1.7
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.13.2
tomli==2.0.1
toolz==0.12.0
torch==1.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
traitlets==5.8.0
transformers @ git+https://github.com/huggingface/transformers.git@60d51ef5123d949fd8c59cd4d3254e711541d278
typer==0.7.0
typing_extensions==4.4.0
uc-micro-py==1.0.1
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.13
uvicorn==0.20.0
wasabi==1.1.1
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
websockets==10.4
xxhash==3.2.0
y-py==0.5.5
yarl==1.8.2
ypy-websocket==0.8.2
zipp==3.13.0
pattern==3.6
